feat: move is_artsy_s3_bucket to validations module
