+++ src/lib/util.py
@@ -34,10 +34,6 @@ def dict1_in_dict2(dict1, dict2):
   ''' return true if all items in dict1 are in dict2 '''
   return set(dict1.items()).issubset(set(dict2.items()))
 
-def is_artsy_s3_bucket(name):
-  ''' return true if bucket name starts with artsy- '''
-  return name.startswith('artsy-')
-
 def is_quoted(str1):
   ''' if string is quoted, return the quote character '''
   # double quote
