+++ src/lib/validations.py
@@ -1,0 +1,3 @@
+def is_artsy_s3_bucket(name):
+  ''' return true if bucket name starts with artsy- '''
+  return name.startswith('artsy-')
