+++ src/rabbitmq_export/export.py
@@ -5,7 +5,7 @@ import os
 import rabbitmq_export.context
 
 from lib.logging import setup_logging
-from lib.util import is_artsy_s3_bucket
+from lib.validations import is_artsy_s3_bucket
 
 from rabbitmq_export.export import (
   export_and_backup
